requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
astroid==2.1.0
Babel==2.6.0
beautifulsoup4==4.7.0
billiard==3.5.0.5
bs4==0.0.1
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
defusedxml==0.5.0
Django==2.1.4
django-admin-sortable2==0.6.21
django-allauth==0.38.0
django-angular==2.2
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.5.3
django-extra-views==0.11.0
django-filer==1.4.1
django-filter==2.0.0
django-formtools==2.1
django-fsm==2.6.0
django-fsm-admin==1.2.4
django-haystack==2.8.1
django-ipware==2.1.0
django-js-asset==1.1.0
django-mptt==0.9.1
django-oscar==1.6.4
django-phonenumber-field==2.0.1
django-polymorphic==2.0.3
django-post-office==3.1.0
django-rest-auth==0.9.3
django-seed==0.1.9
django-sekizai==0.10.0
django-select2==6.3.1
django-shop==0.12.3
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangocms-admin-style==1.2.9
djangocms-cascade==0.17.1
djangocms-text-ckeditor==3.7.0
djangorestframework==3.9.0
djangorestframework-recursive==0.1.2
djangorestframework-timed-auth-token==1.3.0
djangorestframework-xml==1.4.0
djangoshop-paypal==0.2.0
easy-thumbnails==2.5
factory-boy==2.11.1
Faker==1.0.1
html5lib==1.0.1
idna==2.8
isort==4.3.4
jsonfield==2.0.2
kombu==4.2.2.post1
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
oauthlib==2.1.0
paypalrestsdk==1.13.1
pbr==5.1.1
phonenumbers==8.10.2
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
purl==1.4
pycparser==2.19
pyenv==0.1
pylint==2.2.2
pyOpenSSL==18.0.0
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
sorl-thumbnail==12.4.1
soupsieve==1.6.2
stripe==2.17.0
text-unidecode==1.2
typed-ast==1.1.1
Unidecode==1.0.23
urllib3==1.24.1
vine==1.2.0
webencodings==0.5.1
wrapt==1.10.11
xmljson==0.2.0
